chore: code-split of the long (interceptor.ts) file into smaller pieces
